Update OpenSSL to 1.0.1q (fixes #496)

+++ plugins/python-build/test/test_helper.bash
@@ -10,7 +10,7 @@ if [ "$FIXTURE_ROOT" != "$BATS_TEST_DIRNAME/fixtures" ]; then
 fi
 
 teardown() {
-  rm -fr "$TMP"/*
+  rm -fr "${TMP:?}"/*
 }
 
 stub() {
